requirements.txt
json2html
Flask
smbus
RPi.GPIO
